requirements.txt
google-adk[database]==0.3.0
yfinance==0.2.56
psutil==5.9.5
litellm==1.66.3
google-generativeai==0.8.5
python-dotenv==1.1.0
fastapi>=0.115.0
uvicorn>=0.15.0
pydantic>=2.0.0
python-dotenv>=0.19.0
